split egg-info and sdist tests
