+++ tests/test_converters/test_egginfo.py
@@ -8,21 +8,11 @@ from dephell.models import Requirement
 
 
 def test_load_deps():
-    path = Path('tests') / 'requirements' / 'sdist.tar.gz'
-    root = SDistConverter().load(path)
+    path = Path('tests') / 'requirements' / 'egg-info'
+    root = EggInfoConverter().load(path)
 
-    needed = {'attrs', 'cached-property', 'packaging', 'requests'}
+    needed = {'attrs', 'cached-property', 'packaging', 'requests', 'colorama', 'libtest'}
     assert set(dep.name for dep in root.dependencies) == needed
-
-
-def test_load_metadata():
-    path = Path('tests') / 'requirements' / 'sdist.tar.gz'
-    root = SDistConverter().load(path)
-
-    assert root.name == 'dephell'
-    assert root.version == '0.2.0'
-    assert root.authors[0].name == 'orsinium'
-    assert not root.license
 
 
 def test_dumps_deps():
